updated functional tests with fixtures
