+++ cytosnake/tests/functional/test_cli.py
@@ -1,102 +1,155 @@
 """
 module: test_cli.py
 
-This testing module composes of functional tests that contains checks for both positive
-negative cases when using CytoSnake's CLI
+Description:
 
-A positive case indicates that given the user parameters we expect it to run
-successfully.
+The test_cli.py module is a crucial part of the CytoSnake project, responsible for
+conducting functional tests on CytoSnake's Command Line Interface (CLI). These tests aim
+to ensure that the CLI functions correctly, handling both positive and negative
+scenarios effectively.
 
-A negative case indicates that given with the user parameters, our tests are able to
-capture the errors.
+The purpose of this module is to verify that the CytoSnake CLI operates as expected and
+provides accurate results when interacting with various user parameters and modes. It
+validates that the CLI can handle different inputs, execute successfully in positive
+cases, and properly report errors in negative cases.
 
-Ultimately, test_cli.py will contains functional test to all modes that CytoSnake
-contains.
+The test scenarios are categorized into two main types:
+
+Positive Cases:
+These test cases validate the expected behavior of the CLI when
+provided with valid user parameters. The primary goal is to ensure that the CLI executes
+successfully without errors and produces the correct output.
+
+Negative Cases:
+These test cases simulate scenarios where invalid user parameters are provided. 
+The objective is to confirm that the CLI can identify and handle errors appropriately, 
+providing informative error messages to the user.
 """
 
 import os
+from dataclasses import dataclass
+import pytest
 import pathlib
 import shutil
 import subprocess
-from typing import Optional
 
 from cytosnake.common import errors
 
 
-# -----------------
-# Helper functions
-# -----------------
-class CleanUpHandler:
-    """Used to clean up directories in every single test run"""
+# --------------
+# helper functions
+# --------------
+@dataclass
+class DataFiles:
+    """Structured datatype representation that contains all files in a selected dataset
 
-    def __init__(self, tmp_path):
-        self.tmp_path = tmp_path
+    Attributes:
+    -----------
+    metdata: str | list[str]
+        metadata directory name
+    plate_data: list[str]
+        list of plate data (parquet or sqlite files)
+    barcode : str
+        barcode file name
 
-    def __call__(self) -> None:
-        shutil.rmtree(self.tmp_path)
+    Returns
+    -------
+    DataFiles
+        DataStructure representation of test dataset files
+    """
+
+    # required parameters
+    dataset_dir: str | pathlib.Path
+
+    # extracted files
+    metadata: str | list[str] = None
+    plate_data: str | list[str] = None
+    barcode: str = None
+
+    # extracting file paths and setting into dataclass attributes
+    def __post_init__(self):
+        self._extract_content_files()
+
+    def _extract_content_files(self):
+        """extracts all files within given dataset folder and sets the DataFile dataclass
+        attributes
+
+        Raises
+        ------
+        TypeError
+            raised if dataset_dir is not a str or pathlib.Path object.
+            raised if plate data is not parquet or sqlite file
+        """
+        # get all top level files
+        if not isinstance(self.dataset_dir, (str, pathlib.Path)):
+            raise TypeError(
+                "dataset_dir must be a string or pathlib.Path object"
+            )
+        if isinstance(self.dataset_dir, str):
+            self.dataset_dir = pathlib.Path(self.dataset_dir).resolve(
+                strict=True
+            )
+
+        # get all files
+        all_files = list(self.dataset_dir.glob("*"))
+
+        # get data files
+        plate_data = [
+            str(fpath.name)
+            for fpath in all_files
+            if fpath.suffix == ".parquet" or fpath.suffix == ".sqlite"
+        ]
+        self.plate_data = plate_data
+
+        # get metadata_dir
+        meta_data_path = [
+            str(fpath.name) for fpath in all_files if fpath.is_dir()
+        ]
+        self.metadata = (
+            meta_data_path[0] if len(meta_data_path) == 1 else meta_data_path
+        )
+
+        # get barcode
+        barcode_path = [
+            str(fpath.name) for fpath in all_files if fpath.suffix == ".txt"
+        ]
+        self.barcode = (
+            barcode_path if len(barcode_path) == 1 else barcode_path
+        )
 
 
-def transfer_data(
-    test_dir: pathlib.Path,
-    n_plates: int,
-    n_platemaps: int,
-    metadata_dir_name: Optional[str] = "metadata",
-    testing_data_dir="emptyfiles",
-) -> None:
-    """Wrapper function that transfer datasets found within the pytest module and
-    transfers it to the assigned directory where pytest is conducting the functional
-    tests.
-
-    The test dataset is transfered from the `dataset` directory located in the
-    same directory as the functional test module. The dataset directory contains
-    empty files that emulates inputs that a user will put in to CytoSnake.
+def get_test_data_folder(test_data_name: str) -> DataFiles:
+    """Gets single or multiple datasets. Users provide the name of the datasets
+    that will be used in their tests
 
     Parameters
     ----------
-    test_dir : pathlib.Path
-        testing directory
+    test_data_name : str | list[str]
+        name or names of datasets to be selected
 
-    n_plates: int
-        number of plates
+    Returns
+    -------
+    DataFiles
+        contains all files in a dataclass format
 
-    n_platesmaps: int
-        number of plate maps
-
-    metadata_dir_name : Optional[str]
-        name of the metadata directory (default=metadata)
-
-    Return
-    ------
-    None
-        Transfers datafiles from the PyTest module to the testing directory
+    Raises:
+    -------
+    FileNotFoundError
+        Raised when the provided test_data_name is not a valid testing dataset.
     """
 
-    # get files to transfer
-    # dataset folder is within the same directory as the test modules, changes
-    # in this path will cause the tests to fail
-    dataset_dir = pathlib.Path(f"./datasets/{testing_data_dir}").resolve(strict=True)
+    # type checking
+    if not isinstance(test_data_name, str):
+        raise TypeError("`test_data_name` must be a string")
 
-    # grabbing all input paths
-    sqlite_file_paths = list(dataset_dir.glob("*sqlite"))[:n_plates]
-    platemaps_dir = dataset_dir / "metadata" / "platemap"
-    plate_map_files = [
-        str(_path.absolute()) for _path in platemaps_dir.glob("platemap*")
-    ][:n_platemaps]
-    barcode = dataset_dir / "barcode.txt"
+    # get testing dataset_path
+    data_dir_path = pathlib.Path("./datasets")
+    sel_test_data = (data_dir_path / test_data_name).resolve(strict=True)
 
-    # create a metadata_dir in tmp_dir
-    if not isinstance(metadata_dir_name, str):
-        raise ValueError("metadata dir name must be a string")
+    # convert to DataFiles content
+    data_files = DataFiles(sel_test_data)
 
-    tmpdir_metadata_path = test_dir / metadata_dir_name / "platemap"
-    tmpdir_metadata_path.mkdir(exist_ok=True, parents=True)
-
-    # transferring all files to tmp dir
-    for _path in plate_map_files:
-        shutil.copy(_path, str(tmpdir_metadata_path))
-    for _path in sqlite_file_paths:
-        shutil.copy(_path, test_dir)
-    shutil.copy(barcode, test_dir)
+    return data_files
 
 
 def get_raised_error(traceback: str) -> str:
@@ -125,102 +178,178 @@ def get_raised_error(traceback: str) -> str:
     return traceback.splitlines()[-1].split(":")[0].split(".")[-1]
 
 
-# --------------------------
-# init mode functional tests
-# --------------------------
-# The tests below focuses on only executing the init mode.
-def test_barcode_logic_no_barcode_one_platemap(tmp_path, request) -> None:
-    """Positive case: This tests expects a successful run where the user provides
-    multiple plate datasets, plate map, and no barcode. Since this is only one plate_map
-    , this means that the generated dataset came from one experiment and multiple
-    samples (plates) were used to generated the datasets.
+def prepare_dataset(
+    test_data_name: str,
+    test_dir_path: pathlib.Path,
+) -> DataFiles:
+    """Main function to prepare dataset into testing datafolder.
+
+    Parameters
+    ----------
+    test_data_name : str
+        name of the testing dataset you want to use
+
+    test_dir_path : pathlib.Path
+        path to testing directory
+
+    Returns
+    -------
+    DataFiles
+        Structured data object that contains all the files within the selected dataset
     """
-    # starting path
-    test_module = str(pathlib.Path().absolute())
+    # get dataset and transfer to testing directory
+    datafiles = get_test_data_folder(test_data_name=test_data_name)
+    shutil.copytree(datafiles.dataset_dir, str(test_dir_path), dirs_exist_ok=True)
 
-    # transfer placeholder data to tmpdir
-    transfer_data(test_dir=tmp_path, n_plates=2, n_platemaps=1)
+    # change directory to the testing directory
+    os.chdir(str(test_dir_path))
 
-    # change directory to tmpdir
-    os.chdir(tmp_path)
+    return datafiles
 
-    # execute CytoSnake
+
+# ---------------
+# PyTest Fixtures
+# ---------------
+@pytest.fixture
+def testing_dir(tmp_path, request):
+    """Creates a testing directory
+
+    Note: Pytest will tear down tmp_path per test
+
+    Parameters
+    ----------
+    tmp_path : pytest.fixture
+        pytest deafault fixture value to be called when creating a temp dir.
+
+    request : pytest.fixture
+        Allows custom testing function to be added in the testing workflow
+
+    returns:
+    --------
+    pathlib.Path
+        Path pointing to temporary directory
+    """
+    # setting paths
+    original_cwd = str(pathlib.Path(".").absolute())
+
+    # creating a temporary path
+    test_name = request.node.name
+    tmp_dir = (tmp_path / test_name).absolute()
+    tmp_dir.mkdir()
+
+    # return the temporary directory path
+    yield tmp_dir
+
+    # teardown: remove the testing
+    if os.path.exists(tmp_dir):
+        shutil.rmtree(tmp_dir)
+
+    #  change the current working directory to the oroginal root
+    os.chdir(str(original_cwd))
+
+
+# ---------------
+# Init Mode Tests
+# ---------------
+def test_multiplate_maps_no_barcode(testing_dir) -> None:
+    """Negative case test: Expects `BarcodeMissingError` and Non-zero return code
+
+    This tests checks if the CLI raises an error when a user provides multiple platemaps
+    but no barcodes.
+
+    Rational:
+    ---------
+        CytoSnake relies on a barcode file to accurately map each platemap to its
+        corresponding plate. If users provide multiple platemaps without a barcode,
+        it becomes impossible for CytoSnake to determine the correct association between
+        platemaps and their respective plates.
+
+    Parameters:
+    -----------
+    testing_dir: pytest.fixture
+        Testing directory
+    """
+
+    # transfer data to testing folder
+    prepare_dataset(
+        test_data_name="standard_sqlite_multi", test_dir_path=testing_dir
+    )
+
+    # execute test
     cmd = "cytosnake init -d *.sqlite -m metadata".split()
     proc = subprocess.run(cmd, capture_output=True, text=True, check=False)
 
-    # leave test directory
-    os.chdir(test_module)
+    # Grab raised exception
+    raised_exception = get_raised_error(proc.stderr)
 
-    # clean directory,
-    cleanup_handler = CleanUpHandler(tmp_path)
-    request.addfinalizer(cleanup_handler)
-
-    # checking for success return code
-    assert proc.returncode == 0
+    # assert checking
+    assert proc.returncode == 1
+    assert raised_exception == errors.BarcodeMissingError.__name__
 
 
-def test_barcode_logic_barcode_multi_platemaps(tmp_path, request) -> None:
-    """Positive case: This tests expects a successful run where the user provides
-    multiple plate datasets, multiple plate map, and barcode. Since this is only one
-    plate_map , this means that the generated dataset came from one experiment and
-    multiple samples (plates) were used to generated the datasets.
+def test_one_plate_one_platemap(testing_dir) -> None:
+    """Positive case test: Expects a 0 exit code
+
+    This test checks if the CLI can handle one plate and one plate map without barcode
+    as inputs.
+
+    Rational:
+    ---------
+        As this input involves only one plate, CytoSnake does not require the process
+        of identifying which platemap corresponds to that plate.
+
+    Parameters:
+    -----------
+    test_dir: pytest.fixture
+        Testing directory
     """
-    # PyTest module directory
-    test_module = str(pathlib.Path().absolute())
 
-    # transfer placeholder data to tmpdir
-    transfer_data(test_dir=tmp_path, n_plates=2, n_platemaps=2)
+    # prepare testing files
+    datafiles = prepare_dataset(
+        test_data_name="standard_sqlite_single", test_dir_path=testing_dir
+    )
 
-    # change directory to tmpdir
-    os.chdir(tmp_path)
+    # Selecting one plate and meta data dir
+    plate = datafiles.plate_data[0]
+    metadata = datafiles.metadata
 
-    # execute CytoSnake
-    cmd = "cytosnake init -d *.sqlite -m metadata -b barcode.txt".split()
+    # execute
+    cmd = f"cytosnake init -d {plate} -m {metadata}".split()
     proc = subprocess.run(cmd, capture_output=True, text=True, check=False)
 
-    # leave testing dir
-    os.chdir(test_module)
-
-    # clean directory,
-    cleanup_handler = CleanUpHandler(tmp_path)
-    request.addfinalizer(cleanup_handler)
-
-    # checking for success return code
+    # assert check
     assert proc.returncode == 0
 
+def test_multiplates_with_multi_platemaps(testing_dir):
+    """Positive case test: Expects a 0 exit code
 
-def test_barcode_logic_no_barcode_multi_platemaps(tmp_path, request) -> None:
-    """Negative case: This test expects a failed run where the user provides multiple
-    plate datasets, multiple plate maps (multi-experiments), and no barcode. Since
-    there are plate maps, this indicates that the generated datasets came from multiple
-    experiments.
+    This test checks if the CLI can handle multiple plates and multiple plate maps 
+    with a barcode as inputs.
 
-    Checks:
-    -------
-        non-zero return code
-        BarCodeRequiredError raised
+    Rational:
+    ---------
+        CytoSnake requires input from multiple platemaps and barcodes. To ensure 
+        accurate mapping of plate data, it necessitates the use of a barcode as an 
+        input. The barcode files are used by CytoSnake to correctly associate each 
+        plate map with its corresponding plate data.
+
+    Parameters:
+    -----------
+    test_dir: pytest.fixture
+        Testing directory
     """
-    # PyTest module directory
-    test_module = str(pathlib.Path().absolute())
 
-    # transfer placeholder data to tmpdir
-    transfer_data(test_dir=tmp_path, n_plates=2, n_platemaps=2)
+    # prepare testing files
+    datafiles = prepare_dataset(
+        test_data_name="standard_sqlite_multi", test_dir_path=testing_dir
+    )
 
-    # change directory to tmpdir
-    os.chdir(tmp_path)
+    # Selecting one plate and meta data dir
+    barcode = datafiles.barcode
 
-    # execute CytoSnake
-    cmd = "cytosnake init -d *.sqlite -m metadata"
-    proc = subprocess.run(cmd, shell=True, capture_output=True, text=True, check=False)
-    raised_error = get_raised_error(proc.stderr)
+    # execute
+    cmd = f"cytosnake init -d *plate -m metadata -b {barcode}".split()
+    proc = subprocess.run(cmd, capture_output=True, text=True, check=False)
 
-    # leave testing dir
-    os.chdir(test_module)
-
-    # clean directory,
-    cleanup_handler = CleanUpHandler(tmp_path)
-    request.addfinalizer(cleanup_handler)
-
-    # checking for success return code
-    assert proc.returncode == 1
-    assert raised_error == errors.BarcodeRequiredError.__name__
+    # assert checks
+    assert proc.returncode == 0
